changed time type to uint32_t; ignoring songdata.txt

+++ app/key.c
@@ -7,9 +7,9 @@
 
 /// initilize a note
 // uint8_t key is the key to be played
-// uint16_t time is the time to be played at
+// uint32_t time is the time to be played at
 // uint8_t intensity is the intensity of the key being played
-Note* init_note(uint8_t key, uint16_t time, uint8_t intensity) {
+Note* init_note(uint8_t key, uint32_t time, uint8_t intensity) {
     Note *note = malloc(sizeof(Note));
     note->key = key;
     note->time = time;
